feat(#126): 클래스 작성 서버 api 연결
