+++ src/lib/apis/instructorApi.ts
@@ -4,6 +4,7 @@ import {
   IGetClassDraft,
   IUpdateClassDraft,
 } from '@/types/class';
+import { IInstructorRegister } from '@/types/instructor';
 
 export const getInstructorProfile = async () => {
   const response = await fetch(`${DOMAIN}/api/instructors/myProfile`, {
@@ -32,25 +33,7 @@ export const getCheckNickname = async (nickname: string) => {
   }
 };
 
-interface InstructorRegister {
-  profileImageUrls: string[];
-  nickname: string;
-  email: string;
-  phoneNumber: string;
-  profileCardImageUrl?: string;
-  youtubeUrl?: string;
-  instagramUrl?: string;
-  homepageUrl?: string;
-  affiliation?: string;
-  introduction: string;
-  experience?: string;
-  regions: string[];
-  genres: string[];
-  instagramPostUrls?: string[];
-  etcGenres?: string[];
-}
-
-export const instructorRegister = async (data: InstructorRegister) => {
+export const instructorRegister = async (data: IInstructorRegister) => {
   try {
     const response = await fetch(`${DOMAIN}/api/instructors/register`, {
       method: 'POST',
@@ -73,141 +56,6 @@ export const instructorRegister = async (data: InstructorRegister) => {
     return await response.json();
   } catch (error) {
     console.error('강사 등록 오류', error);
-    throw error;
-  }
-};
-
-export const getClassDraft = async (
-  lectureId: number | string,
-): Promise<IGetClassDraft> => {
-  try {
-    const response = await fetch(
-      `${DOMAIN}/api/class/drafts/getDraft?lectureId=${encodeURIComponent(
-        lectureId,
-      )}`,
-      {
-        method: 'GET',
-        credentials: 'include',
-      },
-    );
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(errorData.message || '');
-    }
-
-    const data = await response.json();
-
-    return data.data.temporaryLecture;
-  } catch (error) {
-    console.error('임시저장 목록 조회 오류', error);
-    throw error;
-  }
-};
-
-export const createClassDraft = async () => {
-  try {
-    const response = await fetch(`${DOMAIN}/api/class/drafts/createDraft`, {
-      method: 'POST',
-      credentials: 'include',
-    });
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(
-        `임시저장 생성 오류: ${errorData.message || ''}, status: ${
-          response.status
-        }`,
-      );
-    }
-
-    const data = await response.json();
-    return data.data;
-  } catch (error) {
-    console.error('임시저장 생성 오류', error);
-    throw error;
-  }
-};
-
-export const getClassDrafts = async (): Promise<IGetClassDrafts[]> => {
-  try {
-    const response = await fetch(`${DOMAIN}/api/class/drafts/getDrafts`, {
-      method: 'GET',
-      credentials: 'include',
-    });
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(
-        `임시저장 목록 조회 오류: ${errorData.message || ''}, status: ${
-          response.status
-        }`,
-      );
-    }
-
-    const data = await response.json();
-
-    return data.data.temporaryLectures;
-  } catch (error) {
-    console.error('임시저장 목록 조회 오류', error);
-    throw error;
-  }
-};
-
-export const deleteClassDrafts = async (
-  lectureId: string | number,
-): Promise<IGetClassDrafts> => {
-  try {
-    const response = await fetch(
-      `${DOMAIN}/api/class/drafts/deleteDraft?lectureId=${encodeURIComponent(
-        lectureId,
-      )}`,
-      {
-        method: 'DELETE',
-        credentials: 'include',
-      },
-    );
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(
-        `임시저장 삭제 오류: ${errorData.message || ''}, status: ${
-          response.status
-        }`,
-      );
-    }
-
-    const data = await response.json();
-
-    return data.data.temporaryLectures;
-  } catch (error) {
-    console.error('임시저장 삭제 오류', error);
-    throw error;
-  }
-};
-
-export const updateClassDraft = async (data: IUpdateClassDraft) => {
-  try {
-    const response = await fetch(`${DOMAIN}/api/class/drafts/updateDraft`, {
-      method: 'PATCH',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(
-        `임시저장 오류: ${errorData.message || ''}, status: ${response.status}`,
-      );
-    }
-
-    const res = await response.json();
-    console.log(res);
-  } catch (error) {
-    console.error('임시저장 오류', error);
     throw error;
   }
 };
