+++ src/store/classCreate.ts
@@ -2,11 +2,11 @@ import create from 'zustand';
 import { IprocessedDraft } from '@/types/class';
 
 interface IUseClassCreateStore {
-  lectureData: IprocessedDraft | null;
-  setLectureData: (data: IprocessedDraft) => void;
+  classData: IprocessedDraft | null;
+  setClassData: (data: IprocessedDraft) => void;
 }
 
 export const useClassCreateStore = create<IUseClassCreateStore>((set) => ({
-  lectureData: null,
-  setLectureData: (data) => set({ lectureData: data }),
+  classData: null,
+  setClassData: (data) => set({ classData: data }),
 }));
