+++ src/types/instructor.d.ts
@@ -50,3 +50,21 @@ export interface InstructorApplyData {
   introduction: string;
   experience: string;
 }
+
+export interface IInstructorRegister {
+  profileImageUrls: string[];
+  nickname: string;
+  email: string;
+  phoneNumber: string;
+  profileCardImageUrl?: string;
+  youtubeUrl?: string;
+  instagramUrl?: string;
+  homepageUrl?: string;
+  affiliation?: string;
+  introduction: string;
+  experience?: string;
+  regions: string[];
+  genres: string[];
+  instagramPostUrls?: string[];
+  etcGenres?: string[];
+}
